# server/machine-watch-swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "MachineWatch",
    "version": "v1"
  },
  "host": "",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/inspection/material/{materialID}/{inspectionType}": {
      "put": {
        "tags": [
          "inspection"
        ],
        "operationId": "inspection_ForceInspection",
        "parameters": [
          {
            "type": "integer",
            "name": "materialID",
            "in": "path",
            "required": true,
            "x-nullable": false,
            "format": "int64"
          },
          {
            "type": "string",
            "name": "inspectionType",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/inspection/next-piece/{inspectionType}": {
      "put": {
        "tags": [
          "inspection"
        ],
        "operationId": "inspection_NextPieceInspection",
        "parameters": [
          {
            "type": "string",
            "name": "inspectionType",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "name": "loc",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PalletLocation"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/inspection/counts": {
      "get": {
        "tags": [
          "inspection"
        ],
        "operationId": "inspection_GetCounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/InspectCount"
              }
            },
            "x-nullable": true
          }
        }
      },
      "put": {
        "tags": [
          "inspection"
        ],
        "operationId": "inspection_SetCounts",
        "parameters": [
          {
            "name": "newCounts",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/InspectCount"
              }
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/inspection/type/{ty}": {
      "get": {
        "tags": [
          "inspection"
        ],
        "operationId": "inspection_GetGlobalInspectionType",
        "parameters": [
          {
            "type": "string",
            "name": "ty",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/InspectionType"
            },
            "x-nullable": true
          }
        }
      },
      "put": {
        "tags": [
          "inspection"
        ],
        "operationId": "inspection_SetGlobalInspection",
        "parameters": [
          {
            "type": "string",
            "name": "ty",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "name": "ity",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InspectionType"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "inspection"
        ],
        "operationId": "inspection_DeleteGlobalInspection",
        "parameters": [
          {
            "type": "string",
            "name": "ty",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
    "PalletLocation": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "loc",
        "group",
        "num"
      ],
      "properties": {
        "loc": {
          "$ref": "#/definitions/PalletLocationEnum"
        },
        "group": {
          "type": "string"
        },
        "num": {
          "type": "integer",
          "format": "int32"
        }
      },
      "allOf": [
        {
          "$ref": "#/definitions/ValueType"
        }
      ]
    },
    "PalletLocationEnum": {
      "type": "string",
      "description": "",
      "x-enumNames": [
        "LoadUnload",
        "Machine",
        "MachineQueue",
        "Buffer",
        "Cart"
      ],
      "enum": [
        "LoadUnload",
        "Machine",
        "MachineQueue",
        "Buffer",
        "Cart"
      ]
    },
    "ValueType": {
      "type": "object",
      "x-abstract": true,
      "additionalProperties": false
    },
    "InspectCount": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Counter",
        "Value",
        "LastUTC"
      ],
      "properties": {
        "Counter": {
          "type": "string"
        },
        "Value": {
          "type": "integer",
          "format": "int32"
        },
        "LastUTC": {
          "type": "string",
          "format": "date-time"
        }
      },
      "allOf": [
        {
          "$ref": "#/definitions/ValueType"
        }
      ]
    },
    "InspectionType": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Overrides",
        "Name",
        "TrackPartName",
        "TrackPalletName",
        "TrackStationName",
        "InspectSingleProcess",
        "DefaultCountToTriggerInspection",
        "DefaultDeadline",
        "DefaultRandomFreq"
      ],
      "properties": {
        "Overrides": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InspectionFrequencyOverride"
          }
        },
        "Name": {
          "type": "string"
        },
        "TrackPartName": {
          "type": "boolean"
        },
        "TrackPalletName": {
          "type": "boolean"
        },
        "TrackStationName": {
          "type": "boolean"
        },
        "InspectSingleProcess": {
          "type": "integer",
          "format": "int32"
        },
        "DefaultCountToTriggerInspection": {
          "type": "integer",
          "format": "int32"
        },
        "DefaultDeadline": {
          "type": "string",
          "format": "time-span"
        },
        "DefaultRandomFreq": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "InspectionFrequencyOverride": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Part",
        "CountBeforeInspection",
        "Deadline",
        "RandomFreq"
      ],
      "properties": {
        "Part": {
          "type": "string"
        },
        "CountBeforeInspection": {
          "type": "integer",
          "format": "int32"
        },
        "Deadline": {
          "type": "string",
          "format": "time-span"
        },
        "RandomFreq": {
          "type": "number",
          "format": "double"
        }
      }
    },
...
  },
...
  "securityDefinitions": {}
}